% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceMeasures.R
\name{performanceMeasures}
\alias{performanceMeasures}
\title{performanceMeasures}
\usage{
performanceMeasures(fcst, obs)
}
\arguments{
\item{fcst}{\code{n x m} matrix of \code{n} forecasts with \code{m} ensemble members}

\item{obs}{vector of \code{n} verifying observations}
}
\description{
Performance measures of joint forecast observation distribution
}
\details{
Six key performance measures of the joint probabilitiy distribution of
probabilistic forecasts (re-forecasts) and corresponding observations
are computed for diagnosis of forecast performance.
}
