% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNRparam.R
\name{SNRparam}
\alias{SNRparam}
\title{SNRparam}
\usage{
SNRparam(x)
}
\arguments{
\item{x}{vector with 6 performance measures}
}
\description{
Parameters of signal-to-noise model from population moments
}
\details{
Convert the six key performance measures to parameters of the
  corresponding signal-to-noise model.
}
