% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNRresample.R
\name{SNRresample}
\alias{SNRresample}
\alias{SNRgen}
\alias{SNRgen}
\title{Resample verification statistics}
\usage{
SNRresample(fcst, obs, fun, nboot = 999, mc.cores = 1, simplify = TRUE)

SNRgen(i, fcst, obs)
}
\arguments{
\item{fcst}{\code{n x m} matrix of \code{n} forecasts with \code{m} ensemble
members}

\item{obs}{vector of \code{n} verifying observations}

\item{fun}{One of function, list of functions, or character vector with names
of verification functions to be used with
\code{\link[easyVerification]{veriApply}}}

\item{nboot}{Number of bootstrap replicates to be drawn}

\item{mc.cores}{Number of cores to use for parallel computation}

\item{simplify}{logical or character string; attempt to reduce the result to
a vector, matrix or higher dimensional array; see the simplify argument of
\code{\link[base]{sapply}}.}

\item{i}{number of forecast-obs pair to generate, in case this is set to
zero, the original forecasts and observations are returned}
}
\description{
Using parametric bootstrapping, resample a collection of verification
statistics on the forecasts and observations based on the characterization by
key measures (parameters of the signal-to-noise model) with \code{SNRresample}
and generate bootstrap samples with \code{SNRgen}.
}
\examples{
tm <- easyVerification::toymodel()
summary(SNRresample(tm$fcst, tm$obs, performanceMeasures))

summary(SNRresample(tm$fcst, tm$obs,
                    list(function(fcst, obs) {
                             list(mean_x = mean(fcst), mean_y = mean(obs))
                             },
                         function(fcst, obs) SNRparam(performanceMeasures(fcst, obs)))))

summary(SNRresample(tm$fcst, tm$obs, c("EnsCorr", "Ens2AFC", "FairCrpss")))

}
