% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SNRsample.R
\name{summary.SNRsample}
\alias{summary.SNRsample}
\title{Summary of bootstrapped SNR samples}
\usage{
\method{summary}{SNRsample}(object, plev = c(0.025, 0.975), ...)
}
\arguments{
\item{object}{object of class \code{SNRsample} returned by \code{\link{SNRresample}}}

\item{plev}{percentiles of the bootstrap distribution to be returned}

\item{...}{additional parameters for consistency with generic}
}
\description{
Computes and prints a formatted summary of bootstrap replicates
generated with \code{\link{SNRresample}}.
}
